feat: add websocket streaming support to ux auditor
Implemented real-time AI analysis streaming for the UX Auditor feature. This includes:
- A new AIWebSocketClient with 15s timeouts and exponential backoff.
- Integration of jsonrepair for robust handling of partial JSON streaming chunks.
- UI updates to display live streaming status and progressive analysis results.
- A mock WebSocket implementation for local development and testing.
