+++ package.json
@@ -28,6 +28,7 @@
     "clsx": "^2.1.1",
     "firebase": "^12.12.1",
     "fuse.js": "^7.3.0",
+    "jsonrepair": "^3.14.0",
     "lucide-react": "^0.546.0",
     "motion": "^12.23.24",
     "path-to-regexp": "^8.4.2",
